requirements.txt
hyperliquid-python-sdk==0.3.0
discord==2.3.2
discord.py==2.3.2
ccxt==4.1.10
requests==2.31.0
pandas==2.2.2
numpy==1.23.2
pandas_ta==0.3.14b0
memory-profiler==0.61.0
